Improve extra data ergonomics
Further invesigation maes it seem like there is only one per sequence,
so that's nice.  We can just return the data directly.

Convincing LLMs to write this in a reasonable way is a real chore.  And
I still don't don't really like the output very much.  But I do have a
useable library.

Increment to 0.2.1